requirements.txt
requests
pyyaml
numpy
scikit-learn
pytest